+++ tests/Serilog.Sinks.File.Encrypt.Cli.Tests/Usings.cs
@@ -4,6 +4,7 @@ global using System.Security.Cryptography;
 global using System.Text;
 global using NSubstitute;
 global using Serilog.Sinks.File.Encrypt.Cli.Commands;
+global using Serilog.Sinks.File.Encrypt.Cli.Infrastructure;
 global using Shouldly;
 global using Spectre.Console.Cli;
 global using Spectre.Console.Testing;